dom/base/test/test_clipboard_nbsp.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=359303
https://bugzilla.mozilla.org/show_bug.cgi?id=1769534
-->
<head>
  <meta charset="utf-8" />
  <title>Test for copying non-breaking spaces to the clipboard</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=359303">Mozilla Bug 359303</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1769534">Mozilla Bug 1769534</a>

<p id="display"></p>

<div id="content">
<!-- In a plain-text editable control (such as a textarea or textinput), copying to clipboard should
preserve all non-breaking spaces. -->
<input
  id="input-with-non-breaking-spaces"
  value="Input content: This town is 100&nbsp;km away / «&nbsp;Est-ce Paris&nbsp;?&nbsp;» / Consecutive non-breaking spaces: '&nbsp;&nbsp;' / &nbsp; &nbsp; &nbsp;Text padded using editor-generated pattern.
  ">
<textarea id="textarea-with-non-breaking-spaces">
Textarea content:
- This town is 100&nbsp;km away.
- «&nbsp;Est-ce Paris&nbsp;?&nbsp;»
- Consecutive non-breaking spaces: '&nbsp;&nbsp;'
 &nbsp; &nbsp; &nbsp;Text padded using editor-generated pattern.
</textarea>

<!-- In a content-editable div, copying to clipboard should preserve non-breaking spaces.
However, HTML editors programmatically insert a alternating pattern of space / NBSP when padding text with spaces
(otherwises spaces would be collapsed).
These programmatically-generated NBSPs should be replaced by plain spaces when copying to plain-text.
See https://bugzilla.mozilla.org/show_bug.cgi?id=1769534
-->
<div contenteditable="true" id="content-editable-with-non-breaking-spaces">
Content-editable content:
- This town is 100&nbsp;km away.
- «&nbsp;Est-ce Paris&nbsp;?&nbsp;»
- Consecutive non-breaking spaces: '&nbsp;&nbsp;'
 &nbsp; &nbsp; &nbsp;Text padded using editor-generated pattern.
</div>

<!-- In non-editable HTML nodes, like this paragraph, copying to clipboard should preserve non-breaking spaces.
However, HTML editors programmatically insert a alternating pattern of space / NBSP when padding text with spaces
(otherwises spaces would be collapsed).
These programmatically-generated NBSPs should be replaced by plain spaces when copying to plain-text.
See https://bugzilla.mozilla.org/show_bug.cgi?id=1769534
-->
<p id="paragraph-with-non-breaking-spaces">
Paragraph content:
- This town is 100&nbsp;km away.
- «&nbsp;Est-ce Paris&nbsp;?&nbsp;»
- Consecutive non-breaking spaces: '&nbsp;&nbsp;'
 &nbsp; &nbsp; &nbsp;Text padded using editor-generated pattern.
</p>
</div>

<pre id="test">
<script class="testbody" type="application/javascript">

// Helper: for the given element, select all its content, execute a "Copy" command,
// and return the text put into the clipboard.
async function clipboardTextForElementId(aDomId, aExpectedString) {
  let textContainer = document.getElementById(aDomId);
  let sel = window.getSelection();
  sel.removeAllRanges();

  if (textContainer.select) {
    // Select the entire text in the input or textarea
    textContainer.select();
  } else {
    // Select text node in element.
    let r = document.createRange();
    r.setStart(textContainer, 0);
    r.setEnd(textContainer, 1);
    sel.addRange(r);
  }

  let copiedText = await SimpleTest.promiseClipboardChange(
    function compare(aValue) {
      return aValue.includes(aExpectedString);
    },
    function setup() {
      synthesizeKey("C", {accelKey: true});
    },
    "text/unicode");
  return copiedText;
}

function assertUserNbspPreserved(aCopiedText) {
  ok(aCopiedText.includes("100 km"), "NBSP between two characters should be preserved");
  ok(aCopiedText.includes("« "),  "A single NBSP near a punctuation mark should be preserved");
  ok(aCopiedText.includes(" »"),  "A single NBSP near a punctuation mark should be preserved");
  ok(aCopiedText.includes(" ? "), "NBSPs before *and* after a character should be preserved");
  ok(aCopiedText.includes("'  '"), "Consecutive NBSPs should be preserved");
}

function assertEditorGeneratedNbspPreserved(aCopiedText) {
  ok(aCopiedText.includes("      Text padded"), "When copying from a plain-text control, editor-generated NBSP (if any) should be preserved");
}

function assertEditorGeneratedNbspReplaced(aCopiedText) {
  ok(aCopiedText.includes("      Text padded"), "When copying plain-text from a non plain-text control, editor-generated NBSP should be replaced by standard spaces");
}

/** Test for Bug 359303 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(async function() {
  let textCopiedFromInput = await clipboardTextForElementId("input-with-non-breaking-spaces", "Input");
  assertUserNbspPreserved(textCopiedFromInput);
  assertEditorGeneratedNbspPreserved(textCopiedFromInput);

  let textCopiedFromTextarea = await clipboardTextForElementId("textarea-with-non-breaking-spaces", "Textarea");
  assertUserNbspPreserved(textCopiedFromTextarea);
  assertEditorGeneratedNbspPreserved(textCopiedFromTextarea);

  let textCopiedFromContentEditable = await clipboardTextForElementId("content-editable-with-non-breaking-spaces", "Content-editable");
  assertUserNbspPreserved(textCopiedFromContentEditable);
  assertEditorGeneratedNbspReplaced(textCopiedFromContentEditable);

  let textCopiedFromNonEditableHtmlContent = await clipboardTextForElementId("paragraph-with-non-breaking-spaces", "Paragraph");
  assertUserNbspPreserved(textCopiedFromNonEditableHtmlContent);
  assertEditorGeneratedNbspReplaced(textCopiedFromNonEditableHtmlContent);

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
